Move MobileNavigationButton to separate folder
