+++ src/pages/Dashboard/Container/MobileNavigationButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MobileNavigationButton";
